+++ app/models/annotation.rb
@@ -9,4 +9,7 @@
 #
 class Annotation < ActiveRecord::Base
   validates :name, :presence => true, :uniqueness => true
+
+  has_and_belongs_to_many :documents
+  has_and_belongs_to_many :users
 end