+++ app/models/document.rb
@@ -11,4 +11,7 @@
 #
 class Document < ActiveRecord::Base
   validates :url, :presence => true
+
+  has_and_belongs_to_many :users
+  has_and_belongs_to_many :annotations
 end