+++ CHANGELOG.md
@@ -4,3 +4,9 @@ Features:
   - Add Document model
   - Add Annotation model
   - Add User model
+
+## 0.1.2 (Wed Aug 10, 2016)
+Features:
+  - Add an association (n-n) between Document and Annotation
+  - Add an association (n-n) between Document and User
+  - Add an association (n-n) between Annotation and User